// libs/features/occupational-hazard/src/lib/shared/enums/navigation-indicator.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum NavigationIndicator {
  CSR_EDITS = 1,
  CSR_MODIFY = 7,
  VALIDATOR1_EDITS = 2,
  VALIDATOR1_EDITS_COMPLICATION = 8,
  EST_ADMIN_REOPEN_INJURY = 3,
  EST_ADMIN_REOPEN_GROUP_INJURY = 3,
  CSR_REOPEN_INJURY = 4,
  CSR_REOPEN_GROUP_INJURY = 4,
  ADD_INJURY = 0,
  ADD_GROUP_INJURY=0,
  ADD_DISEASE = 0,
  ADD_OH = 0,
  VALIDATOR1_REOPEN_EDITS = 6,
  VALIDATOR1_EDITS_REPORT = 13,
  REOPEN_COMPLICATION_EST_ADMIN = 9,
  REOPEN_COMPLICATION_CSR = 10,
  REOPEN_COMPLICATION_EDIT = 12,
  ADMIN_UPDATE_PAYEE = 14,
  CSR_UPDATE_PAYEE = 15,
  VALIDATOR1_UPDATE_PAYEE = 16,
  RESUME_CSR_EDITS = 17,
  VALIDATOR_EDIT_DISEASE_DETAILS = 18,
  TRANSFER_INJURY_POST_DISEASE_DETAILS = 19,
  EST_ADMIN_REOPEN_DISEASE_DETAILS_SUBMIT = 21,
  CSR_REOPEN_DISEASE_DETAILS_SUBMIT = 22
}